[Refactor] Move runtime env code under runtime
