+++ be/src/exprs/agg/java_udaf_function.cpp
@@ -23,6 +23,7 @@
 #include "fmt/core.h"
 #include "jni.h"
 #include "platform/user_function_cache.h"
+#include "runtime/java/java_env.h"
 
 namespace starrocks {
