+++ be/src/runtime/global_thread_pools.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "runtime/env/global_thread_pools.h"
+#include "runtime/global_thread_pools.h"
 
 #include <algorithm>
 #include <limits>
