+++ be/src/udf/java/java_data_converter.h
@@ -14,15 +14,19 @@
 
 #pragma once
 
+#include <vector>
+
 #include "column/binary_column.h"
 #include "column/column_visitor.h"
 #include "column/column_visitor_adapter.h"
 #include "column/fixed_length_column.h"
 #include "common/status.h"
 #include "common/statusor.h"
+#include "exprs/function_context.h"
+#include "jni.h"
+#include "runtime/java/java_global_ref.h"
 #include "types/logical_type.h"
 #include "types/type_descriptor.h"
-#include "udf/java/java_udf.h"
 
 namespace starrocks {
 struct JavaUDAFState {
